One more warning to g++ in the Makefile, and changes to fix the errors it caused.

+++ tests/rho/algo/tFFTTest.cpp
@@ -84,7 +84,7 @@ int main()
 {
     tCrashReporter::init();
 
-    srand(time(0));
+    srand((u32)time(0));
 
     tTest("tFFT test", fftTest, kNumTests);
     tTest("tDFT test", dftTest, kNumTests);
