+++ tests/rho/sync/tThreadLocalTest.cpp
@@ -83,7 +83,7 @@ int main()
 {
     tCrashReporter::init();
 
-    srand(time(0));
+    srand((u32)time(0));
     tTest("Thread Local Storage test", test, kTestIterations);
 
     return 0;
